requirements.txt
requests~=2.28.2
validators~=0.20.0
Pillow~=9.4.0
selenium~=4.8.0